refactor: drop slack wrapper, consume @dcl/slack-component directly
The thin wrapper at src/ports/slack/ added one method
(sendFormattedMessage) that simply mapped to @dcl/slack-component's
sendMessage with a default channel. Every caller was the squid
monitor, so the indirection is pure overhead.

- Delete src/ports/slack/{component.ts,index.ts}. Directory goes away.
- src/components.ts: createSlackComponent is imported from
  @dcl/slack-component and invoked inline with SLACK_BOT_TOKEN.
- src/types.ts: ISlackComponent now comes from @dcl/slack-component.
- src/ports/job/squid-monitor.ts: reads SLACK_CHANNEL itself (with
  the same 'general' fallback) and calls slack.sendMessage({
  channel, ...msg }) at each of the 4 alert sites. The
  SlackMessageBlock type is kept here (and exported) because
  @dcl/slack-component types blocks as `any[]` and we want
  structural safety when building alert blocks.

One intentional behaviour change: @dcl/slack-component throws on
failure, whereas the old wrapper swallowed errors and logged them.
Slack failures will now surface through the monitor's outer catch and
be reported via the job's onError instead of being silently absorbed.

10/10 tests still pass; they now mock sendMessage and assert the
first-argument shape still has .text/.blocks (plus a .channel field
that existing assertions don't care about).
