+++ src/ports/job/squid-monitor.ts
@@ -1,11 +1,39 @@
 import { Network } from '@dcl/schemas'
 import { AppComponents } from '../../types'
-import { SlackMessage } from '../slack/component'
 import { Squid } from '../squids/types'
 import { MOCK_SQUIDS } from './mocks'
 
 export const ETA_CONSIDERED_OUT_OF_SYNC = 100
 export const FIVE_MINUTES = 5 * 60 * 1000
+
+export type SlackMessageBlock = {
+  type: 'section' | 'header' | 'divider' | 'context'
+  text?: {
+    type: 'mrkdwn' | 'plain_text'
+    text: string
+  }
+  fields?: Array<{
+    type: 'mrkdwn' | 'plain_text'
+    text: string
+  }>
+  elements?: Array<{
+    type: 'mrkdwn' | 'plain_text'
+    text: string
+  }>
+  accessory?: {
+    type: 'button'
+    text: {
+      type: 'plain_text'
+      text: string
+    }
+    action_id: string
+  }
+}
+
+type SquidAlertMessage = {
+  text: string
+  blocks?: SlackMessageBlock[]
+}
 
 // State to track when "no metrics" issues were first detected for throttling alerts
 export const noMetricsFirstDetected = new Map<string, number>()
@@ -22,6 +50,7 @@ export async function createSquidMonitor(
   const IS_PRODUCTION = (await config.getString('ENV')) === 'prd'
   const ENV_PREFIX = IS_PRODUCTION ? '[PRD]' : '[DEV]'
   const BASE_URL = IS_PRODUCTION ? 'https://decentraland.org/squid-management-ui' : 'https://decentraland.zone/squid-management-ui'
+  const slackChannel = (await config.getString('SLACK_CHANNEL')) || 'general'
   const logger = logs.getLogger('squid-monitor')
 
   const MOCK_ENABLED = (await config.getString('USE_MOCK_SQUIDS')) === 'true' //  this is for testing locally
@@ -58,7 +87,7 @@ export async function createSquidMonitor(
         }
       }
 
-      const errorSlackMessage: SlackMessage = {
+      const errorSlackMessage: SquidAlertMessage = {
         text: `${ENV_PREFIX} 🚨 ERROR: There was a problem monitoring the Squids`,
         blocks: [
           {
@@ -87,7 +116,7 @@ export async function createSquidMonitor(
         ]
       }
 
-      await slack.sendFormattedMessage(errorSlackMessage)
+      await slack.sendMessage({ channel: slackChannel, ...errorSlackMessage })
     }
   }
 
@@ -135,7 +164,7 @@ export async function createSquidMonitor(
           logger.info(`First detection of no metrics for ${squid.service_name} on ${network}. Will send alert after 5 minutes.`)
         } else if (now - firstDetected >= FIVE_MINUTES) {
           // Issue has persisted for 5 minutes, send the alert
-          const noMetricsMessage: SlackMessage = {
+          const noMetricsMessage: SquidAlertMessage = {
             text: `${ENV_PREFIX} ⚠️ ALERT: No metrics found for Squid '${squid.name}' on network ${network}`,
             blocks: [
               {
@@ -177,7 +206,7 @@ export async function createSquidMonitor(
               }
             ]
           }
-          await slack.sendFormattedMessage(noMetricsMessage)
+          await slack.sendMessage({ channel: slackChannel, ...noMetricsMessage })
 
           // Reset the timestamp to avoid spamming (will only send again after another 5 minutes)
           noMetricsFirstDetected.set(noMetricsKey, now)
@@ -201,7 +230,7 @@ export async function createSquidMonitor(
       if (metrics.sqd_processor_sync_eta_seconds === null || metrics.sqd_processor_sync_eta_seconds === undefined) {
         const squidDetailsUrl = `${BASE_URL}?squid=${squid.service_name}&network=${network}`
 
-        const etaUnavailableMessage: SlackMessage = {
+        const etaUnavailableMessage: SquidAlertMessage = {
           text: `${ENV_PREFIX} ⚠️ ALERT: Cannot read ETA for Squid '${squid.name}' on network ${network}`,
           blocks: [
             {
@@ -250,7 +279,7 @@ export async function createSquidMonitor(
           ]
         }
 
-        await slack.sendFormattedMessage(etaUnavailableMessage)
+        await slack.sendMessage({ channel: slackChannel, ...etaUnavailableMessage })
         continue
       }
 
@@ -258,7 +287,7 @@ export async function createSquidMonitor(
       if (metrics.sqd_processor_sync_eta_seconds > ETA_CONSIDERED_OUT_OF_SYNC) {
         const squidDetailsUrl = `${BASE_URL}?squid=${squid.service_name}&network=${network}`
 
-        const desyncMessage: SlackMessage = {
+        const desyncMessage: SquidAlertMessage = {
           text: `${ENV_PREFIX} ⚠️ ALERT: Squid '${squid.name}' on network ${network} is out of sync`,
           blocks: [
             {
@@ -319,7 +348,7 @@ export async function createSquidMonitor(
           ]
         }
 
-        await slack.sendFormattedMessage(desyncMessage)
+        await slack.sendMessage({ channel: slackChannel, ...desyncMessage })
       }
     }
   }
