--- src/ports/slack/index.ts
@@ -1,1 +1,0 @@
-export * from './component'
